refactor: modularize database security.rs for better maintainability
Break down the large security.rs file (848 lines) into focused modules:

- sql_validation.rs: SQL identifier and query validation
- ssl_config.rs: SSL/TLS configuration management
- credentials.rs: Database credentials with environment support
- connection_security.rs: Security validation and configuration loading
- environment.rs: Environment variable configuration utilities
- utils.rs: Error sanitization and utility functions
- mod.rs: Module re-exports for clean API

This modularization improves code organization and follows the structure
outlined in issue #52. All existing functionality is preserved and
verified through comprehensive test coverage (32 database and security tests).
